Make Loops in Non-Link Functions More Efficient
Nearly every loop in the game is now faster. Link-related files were excluded from this because I don't want to inadvertently affect link timings. All i, j, and k variables that were a 8-bit or 16-bit data type have been replaced with their 32-bit equivalents.

Doing this reduces the shifts emitted so dramatically that it saves 4208 bytes in ROM.

+++ gflib/bg.c
@@ -949,7 +949,7 @@ void CopyRectToBgTilemapBufferRect(u8 bg, const void *src, u8 srcX, u8 srcY, u8 
     u16 screenWidth, screenHeight, screenSize;
     u16 var;
     const void *srcPtr;
-    u16 i, j;
+    u32 i, j;
 
     if (!IsInvalidBg32(bg) && !IsTileMapOutsideWram(bg))
     {
