+++ gflib/string_util.c
@@ -14,7 +14,7 @@ static const u8 sDigits[] = __("0123456789ABCDEF");
 
 u8 *StringCopy10(u8 *dest, const u8 *src)
 {
-    u8 i;
+    u32 i;
     u32 limit = 10;
 
     for (i = 0; i < limit; i++)
@@ -31,7 +31,7 @@ u8 *StringCopy10(u8 *dest, const u8 *src)
 
 u8 *StringGetEnd10(u8 *str)
 {
-    u8 i;
+    u32 i;
     u32 limit = 10;
 
     for (i = 0; i < limit; i++)
@@ -82,7 +82,7 @@ u8 *StringAppend(u8 *dest, const u8 *src)
 
 u8 *StringCopyN(u8 *dest, const u8 *src, u8 n)
 {
-    u16 i;
+    u32 i;
 
     for (i = 0; i < n; i++)
         dest[i] = src[i];
@@ -138,7 +138,7 @@ s32 StringCompareN(const u8 *str1, const u8 *str2, u32 n)
 
 bool8 IsStringLengthAtLeast(const u8 *str, s32 n)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < n; i++)
         if (str[i] && str[i] != EOS)
@@ -262,7 +262,7 @@ u8 *ConvertUIntToDecimalStringN(u8 *dest, u32 value, enum StringConvertMode mode
 u8 *ConvertIntToHexStringN(u8 *dest, s32 value, enum StringConvertMode mode, u8 n)
 {
     enum { WAITING_FOR_NONZERO_DIGIT, WRITING_DIGITS, WRITING_SPACES } state;
-    u8 i;
+    u32 i;
     s32 powerOfSixteen;
     s32 largestPowerOfSixteen = 1;
 
@@ -519,7 +519,7 @@ const u8 *GetExpandedPlaceholder(u32 id)
 
 u8 *StringFill(u8 *dest, u8 c, u16 n)
 {
-    u16 i;
+    u32 i;
 
     for (i = 0; i < n; i++)
         *dest++ = c;
@@ -733,7 +733,7 @@ void ConvertInternationalString(u8 *s, u8 language)
 {
     if (language == LANGUAGE_JAPANESE)
     {
-        u8 i;
+        u32 i;
 
         StripExtCtrlCodes(s);
         i = StringLength(s);
