+++ gflib/text.c
@@ -253,7 +253,7 @@ u16 AddTextPrinterParameterized(u8 windowId, u8 fontId, const u8 *str, u8 x, u8 
 bool16 AddTextPrinter(struct TextPrinterTemplate *printerTemplate, u8 speed, void (*callback)(struct TextPrinterTemplate *, u16))
 {
     int i;
-    u16 j;
+    u32 j;
 
     if (!gFonts)
         return FALSE;
