+++ src/battle_anim_effects_2.c
@@ -2382,7 +2382,7 @@ void AnimTask_SketchDrawMon(u8 taskId)
     struct Task* task = &gTasks[taskId];
     struct ScanlineEffectParams params;
 
-    s16 i;
+    s32 i;
     task->data[0] = GetBattlerYCoordWithElevation(gBattleAnimTarget) + 32;
     task->data[1] = 4;
     task->data[2] = 0;
