+++ src/battle_anim_throw.c
@@ -1764,7 +1764,7 @@ static void PokeBallOpenParticleAnimation_Step2(struct Sprite *sprite)
 
 static void TimerBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1798,7 +1798,7 @@ static void TimerBallOpenParticleAnimation(u8 taskId)
 
 static void DiveBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1833,7 +1833,7 @@ static void DiveBallOpenParticleAnimation(u8 taskId)
 // Also used for Net Ball
 static void SafariBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1868,7 +1868,7 @@ static void SafariBallOpenParticleAnimation(u8 taskId)
 // Also used for Nest Ball
 static void UltraBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1903,7 +1903,7 @@ static void UltraBallOpenParticleAnimation(u8 taskId)
 // Also used for Luxury Ball
 static void GreatBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1959,7 +1959,7 @@ static void FanOutBallOpenParticles_Step1(struct Sprite *sprite)
 
 static void RepeatBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -2001,7 +2001,7 @@ static void RepeatBallOpenParticleAnimation_Step1(struct Sprite *sprite)
 
 static void MasterBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i, j;
+    u32 i, j;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -2047,7 +2047,7 @@ static void MasterBallOpenParticleAnimation(u8 taskId)
 
 static void PremierBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
